+++ CHANGELOG.md
@@ -15,6 +15,7 @@ XXXX-XX-XX, v1.0.0
 - merged PR [#319](https://github.com/n1k0/casperjs/pull/319), fixed [#209](https://github.com/n1k0/casperjs/issues/209) - test duration has been added to XUnit XML result file.
 - `Casper.userAgent()` does not require the instance to be started anymore
 - dubious tests now have dedicated color & styling
+- added hint printing when a possible `casperjs` command call is detected
 
 2012-12-14, v1.0.0-RC6
 ----------------------